DocPet/requirements.txt
tabula_py==2.7.0
XlsxWriter==3.0.9
pdfplumber==0.8.0